semantio/knowledge_base/retriever.py
from typing import List, Dict
import numpy as np

class Retriever:
    def __init__(self, vector_store):
        self.vector_store = vector_store

    def retrieve(self, query: str, k: int = 5) -> List[Dict]:
        """
        Retrieve relevant documents for a given query.

        Args:
            query (str): The query string.
            k (int): Number of documents to retrieve.

        Returns:
            List[Dict]: List of relevant documents.
        """
        # Convert the query to an embedding (dummy implementation for now)
        query_embedding = self._embed_query(query)

        # Search the vector store for similar embeddings
        indices = self.vector_store.search(query_embedding, k=k)

        # Retrieve the documents (dummy implementation for now)
        documents = [{"content": f"Document {i}", "score": 0.9} for i in indices]

        return documents

    def _embed_query(self, query: str) -> np.ndarray:
        """
        Convert a query string to an embedding.

        Args:
            query (str): The query string.

        Returns:
            np.ndarray: The query embedding.
        """
        # Dummy implementation: return a random embedding
        return np.random.rand(768)  # Assuming 768-dimensional embeddings