requirements.txt
openai
anthropic
mistralai
google-genai
faiss-cpu
pydantic
requests
groq
playwright
fastapi
uvicorn
pillow
slowapi
sentence-transformers
fuzzywuzzy
duckduckgo-search
yfinance
selenium
beautifulsoup4
webdriver-manager
validators